Fixed transaction script module

+++ src/org/eap/dao/businessobject/Product.java
@@ -4,7 +4,7 @@ public class Product
 {
 	public int ProductID;
 	public int SupplierID;
-	public long Price;
+	public double Price;
 	public String ProductName;
 	public String ProductDescription;
 	public Boolean InStock;
